Fix stardict display for XDXF type
In prior versions the data field from stardicts are printed directly.
This prints out a lot of ugly tags when the stardict is of types other
than plaintext. A simple tag stripping processing is implemented in this
commit. In addition, the application will start from now even if server
fails to launch.