Remove call `QCoreApplication.setApplicationName(app_title(True))`
"""
`QCoreApplication.setApplicationName(app_title(True))` in `vocabsieve/main.py:10` needs to be changed to `app_title(False)`
User data is stored in a location which is based on this setting (via QStandardPaths). Currently users will just lose all the data (dictionary db, records, etc) every time there is an update, which is obviously not ideal.
"""