Implement full lookup db lookup for kindle, koread
The entire lookup database for ereaders will now be imported to the
lookup database of the application. This will happen before any user
action. For this to happen in KOReader, the user must select a directory
containing both the books and the KOReader settings folder. If the
selected folder does not contain KOReader settings (and therefore vocab
database), it will be skipped.